+++ package.json
@@ -42,6 +42,7 @@
     "test:typings": "npm run pretest && npm run build:types && NODE_ENV=fast mocha --config ./config/.mocharc.js ./test_tmp/typings.test.js",
     "test:typings:ci": "npm run pretest && npm run build:types && mocha --config ./config/.mocharc.js ./test_tmp/typings.test.js",
     "test:deps": "dependency-check ./package.json ./dist/lib/plugins/replication-graphql/index.js ./dist/lib/plugins/server.js ./dist/lib/plugins/validate-z-schema.js --no-dev --ignore-module util --ignore-module url --ignore-module \"@types/*\"",
+    "test:circular": "madge --circular src/index.ts",
     "test:performance": "npm run pretest && NODE_ENV=fast mocha --config ./config/.mocharc.js ./test_tmp/performance.test.js",
     "couch:start": "docker run -d -p 5984:5984 --rm --name rxdb-couchdb couchdb:2.1.1",
     "couch:stop": "docker rm -f rxdb-couchdb",
@@ -166,6 +167,7 @@
     "karma-opera-launcher": "1.0.0",
     "karma-safari-launcher": "1.0.0",
     "leveldown": "5.4.0",
+    "madge": "3.4.4",
     "memdown": "5.1.0",
     "mocha": "6.2.2",
     "mocha.parallel": "0.15.6",
