+++ test/unit/core.node.ts
@@ -49,7 +49,7 @@ config.parallel('core.node.js', () => {
                     adapter: 'memory',
                     password: 'myLongAndStupidPassword'
                 }),
-                'RxError',
+                Error,
                 'plugin'
             );
         });
